requirements.txt
livekit-agents[openai,silero,turn-detector,deepgram,google,anthropic,cartesia,elevenlabs,rime,playai,groq,gladia,hedra,inworld,google]
livekit-plugins-aws[realtime]
livekit-plugins-noise-cancellation
python-dotenv
requests>=2.32.0
pydantic
flask
pandas
websockets>=11.0.3
rich
mcp
librosa
moondream
langchain-huggingface
langchain-community
langchain-groq
chromadb
sentence-transformers
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
langchain-groq
fastapi
uvicorn
langchain_chroma